Moved ConditionalEvents and Triggers in dedicated folders
