--- Source/GameBaseFramework/Private/Gameplay/ConditionalEvents/GBFConditionalTrigger.cpp
@@ -1,1 +1,0 @@
-#include "Gameplay/ConditionalEvents/GBFConditionalTrigger.h"
