+++ Source/GameBaseFramework/Public/Triggers/Components/GBFTriggerManagerComponent.h
